client/requirements.txt
requests
tcp_latency
python-nmap